// ANGULAR_APP/Hshop/src/app/Models/product.model.ts
export interface Product {
    id: number;
    productId: number;
    productSku: string;
    productName: string;
    productPrice: number;
    productShortName: string;
    productDescription: string;
    createdDate: string;
    deliveryTimeSpan: string;
    productImageUrl: string;
    categoryName: string;
    rating: number;
  }